app/backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
bcrypt
pyjwt
python-multipart
email-validator